+++ src/main/java/open/commons/spring/jdbc/configure/h2/H2ServerOptions.java
@@ -18,26 +18,34 @@
  *
  * This file is generated under this project, "open-commons-spring-jdbc".
  *
- * Date  : 2025. 4. 3. 오전 11:18:17
+ * Date  : 2025. 4. 3. 오전 11:15:55
  *
  * Author: parkjunhong77@gmail.com
  * 
  */
 
-package open.commons.spring.jdbc.config.h2;
+package open.commons.spring.jdbc.configure.h2;
 
-import java.util.List;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Retention(RUNTIME)
+@Target(METHOD)
 /**
+ * H2 서버 구동 옵션용.
  * 
  * @since 2025. 4. 3.
  * @version 0.5.0
  * @author parkjunhong77@gmail.com
  */
-public interface IH2ServerArgs {
+public @interface H2ServerOptions {
 
     /**
-     * H2 서버 설정을 제공합니다. <br>
+     * H2 옵션 이름에 서버 유형 포함 여부. <br>
+     * '서버 유형' 정보는 'H2 옵션 이름' 앞쪽에 위치함. <br>
      * 
      * <pre>
      * [개정이력]
@@ -51,6 +59,22 @@ public interface IH2ServerArgs {
      * @since 2025. 4. 3.
      * @version 0.5.0
      */
-    public List<String> toArguments();
+    public boolean addType() default true;
 
+    /**
+     * H2 서버 유형 정보가 제외된 옵션 이름. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 4. 3.		parkjunhong77@gmail.com			최초 작성
+     * </pre>
+     *
+     * @return
+     *
+     * @since 2025. 4. 3.
+     * @version 0.5.0
+     */
+    public String name();
 }