+++ src/main/java/open/commons/spring/jdbc/configure/h2/package-info.java
@@ -11,4 +11,4 @@
  * @since 2026. 4. 8.
  */
 @org.jspecify.annotations.NullMarked
-package open.commons.spring.jdbc.config.h2;
+package open.commons.spring.jdbc.configure.h2;