+++ src/main/java/open/commons/spring/jdbc/configure/helper/MultipleDataSourceConfig.java
@@ -24,7 +24,7 @@
  * 
  */
 
-package open.commons.spring.jdbc.config;
+package open.commons.spring.jdbc.configure.helper;
 
 import java.util.ArrayList;
 import java.util.List;